# frontend/docusaurus/static/openapi/index.yaml
openapi: 3.1.0
info:
  title: Matcha - OpenAPI 3.1
  version: 1.0.0
  description: |-
    **Welcome to the Matcha Dating App API!**  
    This specification outlines the Matcha API, a service designed to help people find meaningful connections. The API is built to be:
    - **JSend-compliant**: All responses conform to the [JSend specification](https://github.com/omniti-labs/jsend).
    - **Secure and Privacy-Focused**: We prioritize user safety, data privacy, and consent.
    - **Feature-Rich**: Manage profiles, authenticate users, exchange messages, receive notifications, and more.

    **Useful Resources:**
    - [Matcha GitHub Repository](https://github.com/ugoi/Matcha)
    - [OpenAPI Definition on GitHub](https://github.com/ugoi/Matcha/blob/main/backend/openapi/index.yaml)

  termsOfService: https://github.com/ugoi/Matcha
  contact:
    name: Code Crafting Pro
    url: https://github.com/ugoi/Matcha
    email: codecraftingpro@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
...
  /signup:
    post:
      security: []
      tags:
        - auth
      summary: Sign up
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                username:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    type: object
                    properties:
                      data:
                        type: object
                        properties:
                          title:
                            type: string
                          user:
                            $ref: "#/components/schemas/User"
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/signup' \
            --header 'Content-Type: application/x-www-form-urlencoded' \
            --data-urlencode 'firstName=John' \
            --data-urlencode 'username=john_doe' \
            --data-urlencode 'lastName=Doe' \
            --data-urlencode 'email=john.doe@gmail.com' \
            --data-urlencode 'password=Password123...'
        - lang: typescript
          source: |
            const myHeaders = new Headers();
            myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

            const urlencoded = new URLSearchParams();
            urlencoded.append("firstName", "John");
            urlencoded.append("username", "john_doe");
            urlencoded.append("lastName", "Doe");
            urlencoded.append("email", "john.doe@gmail.com");
            urlencoded.append("password", "Password123...");

            const requestOptions: RequestInit = {
              method: "POST",
              headers: myHeaders,
              body: urlencoded,
              redirect: "follow",
            };

            try {
              const response = await fetch(
                "http://localhost:3000/api/signup",
                requestOptions
              );

              const json = await response.json();
              console.log(json);
              const status = json.status;
              if (status === "error") {
                console.error(status)
              } else if (status === "fail") {
                console.log(status)
              } else if (status === "success") {
                console.error(status)
              }
            } catch (error) {
              console.error("Error:", error);
            }
  /login:
    post:
      security: []
      tags:
        - auth
      summary: Login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Authenticated successfully
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    type: object
                    properties:
                      data:
                        type: object
                        properties:
                          title:
                            type: string
                          token:
                            type: string
                            description: JWT token for authenticated sessions
                          user:
                            $ref: "#/components/schemas/User"
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
          - lang: cURL
            source: |
              curl --location 'http://localhost:3000/api/login' \
              --header 'Content-Type: application/x-www-form-urlencoded' \
              --data-urlencode 'username=john_doe' \
              --data-urlencode 'password=Password123...'
          - lang: typescript
            source: |
              const myHeaders = new Headers();
              myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

              const urlencoded = new URLSearchParams();
              urlencoded.append("username", "john_doe");
              urlencoded.append("password", "Password123...");

              const requestOptions: RequestInit = {
                method: "POST",
                headers: myHeaders,
                body: urlencoded,
                redirect: "follow",
              };

              try {
                const response = await fetch("http://localhost:3000/api/login", requestOptions);
                const json = await response.json();
                const status = json.status;
                if (status === "error") {
                  console.error(status, json);
                } else if (status === "fail") {
                  console.log(status, json.data.errors);
                } else if (status === "success") {
                  console.log("Logged in successfully!", json.data);
                }
              } catch (error) {
                console.error("Error:", error);
              }
...
  /check-auth:
    get:
      tags:
        - auth
      summary: Check authentication
      responses:
        "200":
          description: Session is valid
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/text:
              schema:
                type: string
                example: Unauthorized
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/check-auth'
        - lang: typescript
          source: |
            const requestOptions: RequestInit = {
              method: "GET",
              redirect: "follow",
            };

            try {
              const response = await fetch("http://localhost:3000/api/check-auth", requestOptions);
              const json = await response.json();
              const status = json.status;
              if (status === "error") {
                console.error(status, json);
              } else if (status === "fail") {
                console.log(status, json.data.errors);
              } else if (status === "success") {
                console.log("Session is valid!", json.data);
              }
            } catch (error) {
              console.error("Error:", error);
            }
  /reset-password:
    post:
      security: []
      tags:
        - auth
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        "200":
          description: Password reset link sent
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/reset-password' \
            --header 'Content-Type: application/x-www-form-urlencoded' \
            --data-urlencode 'email=john.doe@gmail.com'
        - lang: typescript
          source: |
            const myHeaders = new Headers();
            myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

            const urlencoded = new URLSearchParams();
            urlencoded.append("email", "john.doe@gmail.com");

            const requestOptions: RequestInit = {
              method: "POST",
              headers: myHeaders,
              body: urlencoded,
              redirect: "follow",
            };

            try {
              const response = await fetch(
                "http://localhost:3000/api/reset-password",
                requestOptions
              );

              const json = await response.json();
              console.log(json);
              const status = json.status;
              if (status === "error") {
                console.error(status)
              } else if (status === "fail") {
                console.log(status)
              } else if (status === "success") {
                console.error(status)
              }
            } catch (error) {
              console.error("Error:", error);
            }

    patch:
      security: []
      tags:
        - auth
      summary: Reset password
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/reset-password' \
            --header 'Content-Type: application/x-www-form-urlencoded' \
            --data-urlencode 'token=123456' \
            --data-urlencode 'password=Password123...'
        - lang: typescript
          source: |
            const myHeaders = new Headers();
            myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

            const urlencoded = new URLSearchParams();
            urlencoded.append("token", "123456");
            urlencoded.append("password", "Password123...");

            const requestOptions: RequestInit = {
              method: "PATCH",
              headers: myHeaders,
              body: urlencoded,
              redirect: "follow",
            };

            try {
              const response = await fetch(
                "http://localhost:3000/api/reset-password",
                requestOptions
              );

              const json = await response.json();
              console.log(json);
              const status = json.status;
              if (status === "error") {
                console.error(status)
              } else if (status === "fail") {
                console.log(status)
              } else if (status === "success") {
                console.error(status)
              }
            } catch (error) {
              console.error("Error:", error);
            }
  /resend-verification-email:
    post:
      security: []
      tags:
        - auth
      summary: Resend verification email
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        "200":
          description: Verification email resent
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/resend-verification-email' \
            --header 'Content-Type: application/x-www-form-urlencoded' \
            --data-urlencode 'john.doe@gmail.com'
        - lang: typescript
          source: |
            const myHeaders = new Headers();
            myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

            const urlencoded = new URLSearchParams();
            urlencoded.append("email", "john.doe@gmail.com");

            const requestOptions: RequestInit = {
              method: "POST",
              headers: myHeaders,
              body: urlencoded,
              redirect: "follow",
            };

            try {
              const response = await fetch(
                "http://localhost:3000/api/resend-verification-email",
                requestOptions
              );

              const json = await response.json();
              console.log(json);
              const status = json.status;
              if (status === "error") {
                console.error(status)
              } else if (status === "fail") {
                console.log(status)
              } else if (status === "success") {
                console.error(status)
              }
            } catch (error) {
              console.error("Error:", error);
            }
  /verify-email:
    patch:
      security: []
      tags:
        - auth
      summary: Verify email
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                    - $ref: "#/components/schemas/SuccessResponse"
                    title: SuccessResponse
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                  - $ref: "#/components/schemas/FailResponse"
                  - $ref: "#/components/schemas/ErrorResponse"
      x-codeSamples:
        - lang: cURL
          source: |
            curl --location 'http://localhost:3000/api/verify-email' \
            --header 'Content-Type: application/x-www-form-urlencoded' \
            --data-urlencode 'token=123456'
        - lang: typescript
          source: |
            const myHeaders = new Headers();
            myHeaders.append("Content-Type", "application/x-www-form-urlencoded");

            const urlencoded = new URLSearchParams();
            urlencoded.append("token", "123456");

            const requestOptions: RequestInit = {
              method: "PATCH",
              headers: myHeaders,
              body: urlencoded,
              redirect: "follow",
            };

            try {
              const response = await fetch(
                "http://localhost:3000/api/verify-email",
                requestOptions
              );

              const json = await response.json();
              console.log(json);
              const status = json.status;
              if (status === "error") {
                console.error(status)
              } else if (status === "fail") {
                console.log(status)
              } else if (status === "success") {
                console.error(status)
              }
            } catch (error) {
              console.error("Error:", error);
            }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
  schemas:
    FieldError:
      type: object
      properties:
        type:
          type: string
          description: The type of error.
          example: field
        value:
          type: string
          description: The input value that caused the error.
          example: adrgonzax2
        msg:
          type: string
          description: The error message.
          example: username already exists
        path:
          type: string
          description: The field or path associated with the error.
          example: username
        location:
          type: string
          description: The location of the error (e.g., body, query, params).
          example: body
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          description: The status of the response.
        data:
          type: object
          nullable: true
          description: The data payload of the response.
    FailResponse:
      type: object
      properties:
        status:
          type: string
          enum: [fail]
          description: The status of the response.
        data:
          type: object
          nullable: true
          description: The data payload of the response.
          properties:
            title:
              type: string
            errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          description: The status of the response.
        message:
          type: string
          description: The error message.
        data:
          type: object
          nullable: true
          description: The data payload of the response.
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user.
        firstName:
          type: string
          description: First name of the user.
        lastName:
          type: string
          description: Last name of the user.
        username:
          type: string
          description: Username of the user.
        email:
          type: string
          format: email
          description: Email address of the user.
...
security:
  - cookieAuth: []